Fix issues raised in review
Config was exported without validation, so a zero value compiled a matcher
that silently never matched, a Limit above 256 panicked while indexing the
alphabet, and a Limit excluding bytes with no ErrRange returned a nil error.
Reject all three with ErrBadConfig.

The oversize guard multiplied before comparing, so on a 32-bit int the
product overflowed, the check passed, and make panicked instead of
returning ErrTooLarge. Compare before multiplying.

The retained-footprint benchmark subtracted unsigned HeapAlloc samples, so
GC noise shrinking the heap underflowed to a huge number rather than a
small negative one.

The README gave the ASCII range as 1-127, but NUL is accepted, and claimed
enforcement was the only remaining difference from ac, omitting the input
folding documented in the paragraph above it.

+++ acascii/bench_test.go
@@ -57,7 +57,9 @@ func BenchmarkCompileRetained(b *testing.B) {
 			}
 			runtime.GC()
 			runtime.ReadMemStats(&after)
-			b.ReportMetric(float64(after.HeapAlloc-before.HeapAlloc), "retained-B")
+			// signed, so that GC noise shrinking the heap stays negative
+			// instead of underflowing to a huge number
+			b.ReportMetric(float64(int64(after.HeapAlloc)-int64(before.HeapAlloc)), "retained-B")
 			b.ReportMetric(0, "ns/op")
 			runtime.KeepAlive(m)
 		})
